SentimentVault/requirements.txt
# Core ML & Data Processing
torch==2.0.1
transformers==4.30.2
datasets==2.12.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0

# FastAPI & Backend
fastapi==0.100.0
uvicorn==0.23.2
pydantic==2.0.2

# Caching & Database
redis==5.0.0

# Load Testing
locust==2.15.1

# Utilities
python-dotenv==1.0.0
tqdm==4.65.0

# Development & Testing
pytest==7.4.0
pytest-asyncio==0.21.1
black==23.7.0
flake8==6.0.0
